requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
starlette==0.36.3
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
argon2-cffi==23.1.0
pyjwt==2.8.0
pydantic==2.6.1
pydantic-settings==2.2.1
structlog==24.1.0
brotli==1.1.0
